Separate name input from main card
