+++ components/profile/player-card/name-input.module.scss
@@ -1,0 +1,5 @@
+.title {
+  font-size: 50px;
+  letter-spacing: -5px;
+  cursor: pointer;
+}
